@@ -1,76 +1,50 @@
 {
-  "name": "maple-demage-skin-simulator",
-  "version": "0.1.0",
+  "name": "vite-project",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "homepage": "https://0407chan.github.io/maple-demage-skin-simulator",
   "scripts": {
     "switch-email": "git config --global user.email 0407chan@naver.com",
-    "dev": "yarn switch-email && BROWSER=none craco start",
-    "build": "rm -rf docs && GENERATE_SOURCEMAP=false craco build && mv build docs",
-    "test": "react-scripts test",
-    "predeploy": "GENERATE_SOURCEMAP=false craco build",
-    "deploy": "gh-pages -d build"
+    "dev": "pnpm switch-email && vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "gh-deploy": "gh-pages -d build"
   },
   "dependencies": {
-    "@craco/craco": "^6.2.0",
-    "@reduxjs/toolkit": "^1.6.1",
+    "@tanstack/react-query": "^4.28.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/react-redux": "^7.1.9",
-    "antd": "^4.16.13",
+    "antd": "^4.24.8",
     "axios": "^0.21.4",
-    "craco-alias": "^3.0.1",
     "env-cmd": "^10.1.0",
-    "firebase": "^7.21.0",
-    "immutability-helper": "^3.1.1",
-    "node-sass": "^4.14.1",
-    "react": "^16.13.1",
-    "react-dnd": "^14.0.3",
-    "react-dnd-html5-backend": "^14.0.1",
-    "react-dnd-multi-backend": "^6.0.2",
-    "react-dnd-touch-backend": "^14.1.1",
-    "react-dom": "^16.13.1",
-    "react-draggable": "^4.4.3",
-    "react-ga4": "^1.4.1",
-    "react-is": "^17.0.2",
-    "react-query": "^3.32.2",
-    "react-redux": "^7.2.1",
-    "react-scripts": "^4.0.3",
-    "recoil": "^0.5.2",
-    "redux": "^4.0.5",
-    "styled-components": "^5.3.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-ga4": "^2.1.0",
+    "recoil": "^0.7.7",
+    "sass": "^1.59.3",
+    "styled-components": "^5.3.9"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.6.1",
-    "@types/react": "^17.0.18",
-    "@types/react-dnd-multi-backend": "^6.0.1",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-query": "^1.2.9",
-    "@types/styled-components": "^5.1.12",
-    "@types/uuid": "^8.3.1",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "babel-plugin-import": "^1.13.3",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.18",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/styled-components": "^5.1.26",
+    "@types/uuid": "^8.3.4",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "@vitejs/plugin-react-swc": "^3.2.0",
+    "babel-plugin-import": "^1.13.6",
     "eslint": "^7.32.0",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react": "^7.32.2",
     "gh-pages": "^3.2.3",
-    "prettier": "^2.3.2",
-    "typescript": "^4.3.5",
-    "uuid": "^8.3.2"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "packageManager": "yarn@3.0.2"
+    "prettier": "^2.8.6",
+    "typescript": "^4.9.5",
+    "uuid": "^8.3.2",
+    "vite": "^4.2.1",
+    "vite-plugin-checker": "^0.5.6",
+    "vite-tsconfig-paths": "^4.0.7"
+  }
 }
