@@ -1,24 +1,20 @@
 import React from 'react'
 import { CSSProperties } from 'styled-components'
+import { GapOptionType } from '../Horizontal/Horizontal'
 import * as S from './style'
-
-export type GapOptionType = 'small' | 'middle' | 'large' | 'none'
 
 type Props = {
   gap?: number | GapOptionType
   style?: CSSProperties
-  onClick?: () => void
   children?: React.ReactNode
 }
-export default function Horizontal({
-  gap,
-  style,
-  onClick,
-  children
-}: Props): JSX.Element {
+
+const Vertical: React.FC<Props> = ({ gap, style, children }) => {
   return (
-    <S.Container gap={gap} style={style} onClick={onClick}>
+    <S.Container gap={gap} style={style}>
       {children}
     </S.Container>
   )
 }
+
+export default Vertical
